Extracted common fetchWhoAmIWithToken
