+++ src/Data/CurrentUser/CurrentUserDataItem.ts
@@ -12,12 +12,9 @@ import { ensureString } from '../../utils/ensureString'
 import { isOptionalString } from '../../utils/isOptionalString'
 import { neoletterClient } from '../neoletterClient'
 import { getTokenAuthorization } from '../getTokenAuthorization'
-import { errorToast, simpleErrorToast } from './errorToast'
-import { pisaClient, pisaConfig } from '../pisaClient'
-import {
-  localizeExpiredMessage,
-  localizeFailedVerify,
-} from './verifySameWhoAmIUser'
+import { errorToast } from './errorToast'
+import { pisaClient } from '../pisaClient'
+import { fetchWhoAmIWithToken } from './fetchWhoAmIWithToken'
 
 async function attributes(): Promise<DataAttributeDefinitions> {
   const lang = await load(currentLanguage)
@@ -176,28 +173,8 @@ async function pisaIds(): Promise<{
 async function getTokenBasedCurrentUser(tokenAuthorization: string) {
   if (isUserLoggedIn()) return null // Safeguard
 
-  const whoAmIConfig = await pisaConfig('whoami')
-  if (!whoAmIConfig) return null
-
-  const lang = await load(() => currentLanguage() ?? '')
-
-  const { url, headers: baseHeaders } = whoAmIConfig
-
-  const headers = { ...baseHeaders, Authorization: tokenAuthorization }
-
-  // TODO: Replace fetch with pisaClient, once #11616 is resolved
-  const response = await fetch(url, { method: 'GET', headers })
-  if (!response.ok) {
-    const errorMessage =
-      response.status === 401
-        ? localizeExpiredMessage(lang)
-        : localizeFailedVerify(lang)
-    simpleErrorToast(errorMessage)
-
-    throw new DataConnectionError(errorMessage)
-  }
-
-  const whoAmI = (await response.json()) as WhoAmI
+  const whoAmI = await fetchWhoAmIWithToken(tokenAuthorization)
+  if (!whoAmI) return null
 
   return {
     company: '',
